增加创建时间，并修改排序

+++ web2/src/views/studentManage/studentNames.vue
@@ -2,18 +2,18 @@
     <div class="box">
         <el-form ref="studentForm" inline :model="filterForm" :rules="rules" label-width="90px">
             <el-form-item label="学生姓名">
-                <el-input placeholder="请输入学生姓名" v-model="filterForm.name"></el-input>
+                <el-input size="mini" placeholder="请输入学生姓名" v-model="filterForm.name"></el-input>
             </el-form-item>
             <!-- <el-form-item label="学生学号">
                 <el-input placeholder="请输入学生学号"></el-input>
             </el-form-item> -->
             <el-form-item label="身份证号">
-                <el-input placeholder="请输入身份证号" v-model="filterForm.studyID"></el-input>
+                <el-input size="mini" placeholder="请输入身份证号" v-model="filterForm.studyID"></el-input>
             </el-form-item>
             <el-form-item>
-                <el-button type="primary" icon="el-icon-search" @click="searchHandle">查询</el-button>
-                <el-button type="primary" @click="dialogVisible = true">添加学生</el-button>
-                <el-button type="warning">重置</el-button>
+                <el-button size="mini" type="primary" icon="el-icon-search" @click="searchHandle">查询</el-button>
+                <el-button size="mini" type="primary" @click="dialogVisible = true">添加学生</el-button>
+                <el-button size="mini" type="warning">重置</el-button>
             </el-form-item>
         </el-form>
         <div>
@@ -34,8 +34,8 @@
                 <el-table-column prop="address" label="家庭住址" align="center"></el-table-column>
                 <el-table-column label="操作" align="center" width="220">
                  <template slot-scope="scope">
-                    <el-button @click="changeStudent(scope.row)" type="primary">修改</el-button>
-                    <el-button @click="deleteStudent(scope.row)" type="danger">删除</el-button>
+                    <el-button size="mini" @click="changeStudent(scope.row)" type="primary">修改</el-button>
+                    <el-button size="mini" @click="deleteStudent(scope.row)" type="danger">删除</el-button>
                 </template>
                 </el-table-column>
             </el-table>
@@ -54,10 +54,12 @@
         <el-dialog
             title="添加学生"
             :visible.sync="dialogVisible"
+            @close="resetAddForm"
+            :show-close="false"
             width="500px">
             <div class="dialog-content">
-                <el-form ref="studentForm" :model="addStudentForm" label-width="90px">
-                    <el-form-item label="学生姓名">
+                <el-form ref="studentForm" :rules="rules" :model="addStudentForm" label-width="90px">
+                    <el-form-item label="学生姓名" prop="name">
                         <el-input placeholder="请输入学生姓名" v-model="addStudentForm.name"></el-input>
                     </el-form-item>
                     <el-form-item label="身份证号码">
@@ -98,10 +100,43 @@
                 <el-button type="primary" @click="addStudent">确 定</el-button>
             </span>
         </el-dialog>
+
+        <el-dialog
+            title="修改资料"
+            :visible.sync="dialogVisible2"
+            @close="resetStudentInfo"
+            :show-close="false"
+            width="500px">
+            <div class="dialog-content">
+                <el-form ref="studentForm" :model="changeStudentInfo" label-width="100px">
+                    <el-form-item label="学生姓名">
+                        <el-input placeholder="请输入学生姓名" v-model="changeStudentInfo.name"></el-input>
+                    </el-form-item>
+                    <el-form-item label="学生年龄">
+                        <el-input placeholder="请输入学生年龄" v-model="changeStudentInfo.age"></el-input>
+                    </el-form-item>
+                    <el-form-item label="学生父亲姓名">
+                        <el-input placeholder="请输入学生父亲姓名" v-model="changeStudentInfo.fatherName"></el-input>
+                    </el-form-item>
+                    <el-form-item label="学生母亲姓名">
+                        <el-input placeholder="请输入学生母亲姓名" v-model="changeStudentInfo.matherName"></el-input>
+                    </el-form-item>
+                    <el-form-item label="学生家庭住址">
+                        <el-input placeholder="请输入学生家庭住址" v-model="changeStudentInfo.address"></el-input>
+                    </el-form-item>
+                </el-form>
+            </div>
+            <span slot="footer" class="dialog-footer">
+                <el-button @click="dialogVisible2 = false">取 消</el-button>
+                <el-button type="primary" @click="changeSrudentInfo">确 定</el-button>
+            </span>
+        </el-dialog>
+
+
     </div>
 </template>
 <script>
-import { addStudent, getStudentList,deleteStudent } from "@/api/studentManage/index.js"
+import { addStudent, getStudentList,deleteStudent,updateStudent } from "@/api/studentManage/index.js"
 export default {
     name:"studentInfo",
     data() {
@@ -117,9 +152,8 @@ export default {
                 name:"",
                 studyID:"",
             },
-            //表单验证规则
-            rules:{},
             dialogVisible:false,
+            dialogVisible2:false,
             addStudentForm:{
                 name: '',
                 studyID:'',
@@ -134,7 +168,30 @@ export default {
             studentSex:[
                 {label:"男",value:'1'},
                 {label:"女",value:"0"}
-            ]
+            ],
+            changeStudentInfo:{
+                studyID:"",
+                name:"",
+                age:"",
+                fatherName:"",
+                matherName:'',
+                address:'',
+            },
+            //表单验证规则
+            rules:{
+                name:[
+                    {required: true, message: '请输入学生姓名', trigger: 'blur'}
+                ]
+            },
+
+        }
+    },
+    watch:{
+        tableData(newVal){
+            if(newVal.length == 0){
+                this.page.current = this.page.current - 1;
+                this.getStudentList()
+            }
         }
     },
     created() {
@@ -172,26 +229,61 @@ export default {
                }
             })
         },
+        //搜索
         searchHandle(){
             this.page.current = 1;
             this.getStudentList();
         },
+        //改变列表尺寸
         handleSizeChange(size){
             this.page.size = size;
             this.page.current = 1;
             this.getStudentList();
         },
+        //跳转页数
         handleCurrentChange(current){
             this.page.current = current;
             this.getStudentList();
         },
-        changeStudent(){},
-        deleteStudent(row){
-            let data = {studyID:row.studyID};
-            deleteStudent(data).then(res=>{
-                console.log(res)
-                this.getStudentList()
+        //修改学生信息
+        changeStudent(row){
+            Object.assign(this.changeStudentInfo,row)
+            // this.changeStudentInfo.studyID = row.studyID;
+            this.dialogVisible2 = true
+        },
+        changeSrudentInfo(){
+            updateStudent(this.changeStudentInfo).then(res =>{
+                if(res.success){
+                    this.$message.success("修改成功");
+                    this.dialogVisible2 = false;
+                    this.getStudentList();                }
             })
+        },
+        resetStudentInfo(){},
+        //删除学生
+        deleteStudent(row){
+            this.$confirm('此操作将永久删除该学生, 是否继续?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+                }).then(() => {
+                    let data = {studyID:row.studyID};
+                    deleteStudent(data).then(res=>{
+                        if(res.success){
+                            this.$message.success('删除成功')
+                            this.getStudentList()
+                        }
+                    })
+                }).catch(() => {
+                    this.$message.warning('已取消删除');
+            });
+        },
+        //重置添加学生表单信息
+        resetAddForm(){
+            // for(let i in this.addStudentForm){
+            //     this.addStudentForm[i] = ""
+            // }
+            // this.refs['studentForm'].resetFields();
         }
